remove my name from copyright headers
This is of course, kind of dumb to need, but unfortunately there are
patent trolls and crazy people out there.

I think we've established that all of everything that has appeared in
the repo is MIT.

It still needs to be defended against an agressor, though. So making sure
there's a clear entity listed is important, having a name there was good.
However, with contributions from other people, I want it to read more
inclusively. Documenting the update in a commit like so will help trace
it backwards in case there is any dispute.